requirements.txt
Django>=2.0,<2.1
django-countries
django_debug_toolbar
django_extensions
django-filter
django-prettyjson
django-rest-swagger
djangorestframework
djangorestframework-filters
idna
psycopg2-binary
pygments
python-memcached
pytz
requests
requests-futures
uwsgi_tasks
